chore(realtime): make tests pass mypy checks

+++ src/realtime/src/realtime/_async/timer.py
@@ -6,7 +6,7 @@ logger = logging.getLogger(__name__)
 
 
 class AsyncTimer:
-    def __init__(self, callback: Callable, timer_calc: Callable[[int], int]):
+    def __init__(self, callback: Callable, timer_calc: Callable[[int], float]):
         self.callback = callback
         self.timer_calc = timer_calc
         self.timer: Optional[asyncio.Task] = None
